feat: 修改一些变量名和类型

+++ raster-go/tri_test.go
@@ -44,7 +44,7 @@ func BarycentricTriColorTest(x1, y1, x2, y2, x3, y3 int, png *image.RGBA) {
 
 	for x := minX; x < maxX; x++ {
 		for y := minY; y < maxY; y++ {
-			a, b, c := ma.Barycentric(x1, y1, x2, y2, x3, y3, x, y)
+			a, b, c := ma.Barycentric(x1, y1, x2, y2, x3, y3, float64(x), float64(y))
 			// 判断是否在三角形内
 			if a < 0 || b < 0 || c < 0 {
 				continue
